reels_shorts/reels_to_shorts/requirements.txt
google-api-python-client
google-auth<4.0.0
google_auth_oauthlib
instagrapi==1.18.4
moviepy==2.2.1
oauth2client<4.0.0
pydantic==1.10.9
python-dotenv
requests
